% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oc_coci.R
\name{oc_coci}
\alias{oc_coci}
\alias{oc_coci_refs}
\alias{oc_coci_cites}
\alias{oc_coci_meta}
\alias{oc_coci_citation}
\title{COCI: OpenCitations Index of Crossref open
DOI-to-DOI references}
\usage{
oc_coci_refs(doi, ...)

oc_coci_cites(doi, ...)

oc_coci_meta(doi, ...)

oc_coci_citation(oci, ...)
}
\arguments{
\item{doi}{(character) one or more Digital Object Identifiers}

\item{...}{curl options passed on to \link[crul:verb-GET]{crul::verb-GET}}

\item{oci}{(character) a single Open Citation Identifier (OCI)}
}
\value{
data.frame
}
\description{
COCI: OpenCitations Index of Crossref open
DOI-to-DOI references
}
\examples{
doi1 <- "10.1108/jd-12-2013-0166"
doi2 <- "10.6084/m9.figshare.3443876"
oci1 <- "02001010806360107050663080702026306630509-0200101080636102704000806" 

# references
oc_coci_refs(doi1)
oc_coci_refs(doi2)

# citations
oc_coci_cites(doi1)
oc_coci_cites(doi2)

# metadata
oc_coci_meta(doi1)
oc_coci_meta(doi2)
oc_coci_meta(c(doi1, doi2))

# citation - an OCI instead of a DOI
oc_coci_citation(oci1)
}
