% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citecorp-package.R
\docType{data}
\name{oc_pmcids}
\alias{oc_pmcids}
\title{Vector of 8 PMCIDs (PubMed Central Identifiers)}
\description{
To be used in examples, etc. for \link{oc_lookup} and \link{oc_coci}
methods
}
\keyword{data}
