% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citecorp-package.R
\docType{package}
\name{citecorp-package}
\alias{citecorp-package}
\alias{citecorp}
\title{Client for the Open Citations Corpus}
\description{
Client for the Open Citations Corpus
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
