% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citecorp-package.R
\docType{package}
\name{citecorp-package}
\alias{citecorp-package}
\alias{citecorp}
\title{citecorp}
\description{
Client for the Open Citations Corpus http://opencitations.net/
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/citecorp (devel)}
  \item \url{https://docs.ropensci.org/citecorp/ (docs)}
  \item Report bugs at \url{https://github.com/ropensci/citecorp/issues}
}

}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}

David Selby \email{david_antony.selby@dfki.de}
}
\keyword{package}
